Move utils.py to nvfuser.testing
so it can be accessed by tests in a different folder.
