+++ tests/python/opinfo_framework.py
@@ -6,7 +6,7 @@
 import inspect
 import torch
 from typing import Callable
-from utils import map_dtype_to_str
+from nvfuser.testing.utils import map_dtype_to_str
 import pytest
 
 
@@ -70,7 +70,7 @@ class create_op_test:
 # correctness tests to avoid calling FusionCache.reset().
 class atexit_serde_create_op_test(create_op_test):
     def __init__(self, opinfos, *, scope=None):
-        from utils import atexit_serde_check
+        from nvfuser.testing.utils import atexit_serde_check
 
         atexit_serde_check()
