+++ tests/python/test_alias.py
@@ -5,7 +5,7 @@
 import torch
 
 from nvfuser import FusionDefinition
-from utils import NVFuserTest
+from nvfuser.testing.utils import NVFuserTest
 
 
 class TestAlias(NVFuserTest):
